+++ gdbgui/sessionmanager.py
@@ -79,11 +79,17 @@ class SessionManager(object):
     def add_new_debug_session(
         self, *, gdb_command: str, mi_version: str, client_id: str
     ) -> DebugSession:
-        pty_for_gdb = Pty(cmd=gdb_command)
         pty_for_debugged_program = Pty()
         pty_for_gdbgui = Pty(echo=False)
-        pty_for_gdb.write(f"new-ui {mi_version} {pty_for_gdbgui.name}\n")
-        pty_for_gdb.write(f"set inferior-tty {pty_for_debugged_program.name}\n")
+        gdbgui_startup_cmds = [
+            f"new-ui {mi_version} {pty_for_gdbgui.name}",
+            f"set inferior-tty {pty_for_debugged_program.name}",
+        ]
+        # instead of writing to the pty after it starts, add startup
+        # commands to gdb. This allows gdb to be run as sudo and prompt for a
+        # password, for example.
+        gdbgui_startup_cmds_str = " ".join([f"-ex='{c}'" for c in gdbgui_startup_cmds])
+        pty_for_gdb = Pty(cmd=f"{gdb_command} {gdbgui_startup_cmds_str}")
 
         pid = pty_for_gdb.pid
         debug_session = DebugSession(
@@ -129,12 +135,6 @@ class SessionManager(object):
         for debug_session in to_remove:
             self.remove_debug_session(debug_session)
 
-    def get_client_ids_from_gdb_pid(self, pid: int) -> List[str]:
-        debug_session = self.debug_session_from_pid(pid)
-        if debug_session:
-            return self.debug_session_to_client_ids.get(debug_session, [])
-        return []
-
     def get_pid_from_debug_session(self, debug_session: DebugSession) -> Optional[int]:
         if debug_session and debug_session.pid:
             return debug_session.pid
@@ -153,19 +153,11 @@ class SessionManager(object):
                 return debug_session
         return None
 
-    def exit_all_gdb_processes(self):
-        logger.info("exiting all subprocesses")
-        for debug_session in self.debug_session_to_client_ids:
-            # TODO kill gdb process and controller
-            logger.info(f"Exiting debug session for pid {debug_session.pid}")
-            debug_session.terminate()
-            self.debug_session_to_client_ids.pop(debug_session)
-
-    def get_dashboard_data(self) -> List[Any]:
-        data = []
-        for debug_session, client_ids in self.debug_session_to_client_ids.items():
-            data.append(debug_session.to_dict())
-        return data
+    def get_dashboard_data(self) -> List[DebugSession]:
+        return [
+            debug_session.to_dict()
+            for debug_session in self.debug_session_to_client_ids.keys()
+        ]
 
     def disconnect_client(self, client_id: str):
         for debug_session, client_ids in self.debug_session_to_client_ids.items():
@@ -173,9 +165,3 @@ class SessionManager(object):
                 client_ids.remove(client_id)
                 debug_session.remove_client(client_id)
         self.remove_debug_sessions_with_no_clients()
-
-    def _spawn_new_gdb_controller(self):
-        pass
-
-    def _connect_to_existing_gdb_controller(self):
-        pass