+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add ability to specify gdb command from the browser. This can now be accomplished from the dashboard.
 * Removed gdbgui binaries from source control. They can now be downloaded as artifacts of [releases](https://github.com/cs01/gdbgui/releases).
 * [documentation] Fix bug when generating md5 checksum for binary releases
+* Remove "shutdown" button in UI
 
 ## 0.13.2.1
 * No end user changes. This release builds the gdbgui executables with GitHub actions.